refactor: fix import paths and tests
